fix: removed UI dir, Accent color added
