+++ postgres_config.yaml
@@ -1,0 +1,130 @@
+# Data sources to query
+data_sources:
+  - name: iris # dataset
+    type: postgres
+    connection:
+      host: 127.0.0.1
+      port: 5421
+      username: !ENV ${PGSQL_USER}
+      password: !ENV ${PGSQL_PASS}
+      database: dcs_db
+      schema: public
+metrics:
+  # Numeric Metrics
+  - name: postgres_avg_price
+    metric_type: avg
+    resource: iris.dcs_iris.sepal_length
+    validation:
+      threshold: "> 0 & < 1000"
+
+  - name: postgres_min_price
+    metric_type: min
+    resource: iris.dcs_iris.sepal_length
+
+  - name: postgres_max_price
+    metric_type: max
+    resource: iris.dcs_iris.sepal_length
+
+  - name: postgres_variance_price
+    metric_type: variance
+    resource: iris.dcs_iris.sepal_length
+
+  # Uniqueness Metrics
+  - name: postgres_distinct_count_price
+    metric_type: distinct_count
+    resource: iris.dcs_iris.sepal_length
+
+  - name: postgres_duplicate_count_price
+    metric_type: duplicate_count
+    resource: iris.dcs_iris.sepal_length
+
+  # Completeness Metrics
+  - name: postgres_null_count_price
+    metric_type: null_count
+    resource: iris.dcs_iris.sepal_length
+
+  - name: postgres_null_percentage_price
+    metric_type: null_percentage
+    resource: iris.dcs_iris.sepal_length
+
+  - name: postgres_empty_string_count_price
+    metric_type: empty_string_count
+    resource: iris.dcs_iris.species
+
+  # Reliability Metrics
+  - name: postgres_row_count
+    metric_type: row_count
+    resource: iris.dcs_iris
+    validation:
+      threshold: "> 100"
+
+  - name: postgres_freshness
+    metric_type: freshness
+    resource: iris.dcs_iris.timestamp
+    validation:
+      threshold: "<5000"
+
+  # Combined Metrics
+  - name: postgres_combined_div
+    metric_type: combined
+    expression: div(postgres_distinct_count_price, postgres_max_price)
+
+  - name: postgres_combined_mul
+    metric_type: combined
+    expression: mul(postgres_distinct_count_price, postgres_max_price)
+
+  - name: postgres_combined_sum
+    metric_type: combined
+    expression: sum(postgres_distinct_count_price, postgres_max_price)
+
+  - name: postgres_combined_sub
+    metric_type: combined
+    expression: sub(postgres_distinct_count_price, postgres_max_price)
+
+  - name: postgres_combined_div_mul
+    metric_type: combined
+    expression: div(mul(postgres_distinct_count_price, postgres_max_price), postgres_min_price)
+
+  - name: postgres_combined_div_mul_sum
+    metric_type: combined
+    expression: sum(div(mul(postgres_distinct_count_price, postgres_max_price), postgres_min_price), postgres_row_count)
+
+  - name: postgres_combined_div_mul_sum_sub
+    metric_type: combined
+    expression: sub(sum(div(mul(postgres_distinct_count_price, postgres_max_price), postgres_min_price), postgres_row_count), postgres_duplicate_count_price)
+
+  - name: postgres_combined_div_mul_sum_sub_div
+    metric_type: combined
+    expression: div(sub(sum(div(mul(postgres_distinct_count_price, postgres_max_price), postgres_min_price), postgres_row_count), postgres_duplicate_count_price), postgres_distinct_count_price)
+
+  # Filtered Metrics
+
+  - name: postgres_avg_price_filtered
+    metric_type: avg
+    resource: iris.dcs_iris.sepal_length
+    filters:
+      where: "species = 'versicolor'"
+
+  - name: postgres_min_price_filtered
+    metric_type: min
+    resource: iris.dcs_iris.sepal_length
+    filters:
+      where: "species = 'versicolor'"
+
+  - name: postgres_max_price_with_filter
+    metric_type: max
+    resource: iris.dcs_iris.sepal_length
+    filters:
+      where: "species = 'versicolor'"
+
+  - name: postgres_min_price_with_filter
+    metric_type: min
+    resource: iris.dcs_iris.sepal_length
+    filters:
+      where: "species = 'versicolor'"
+
+  - name: postgres_distinct_count_price_with_filter
+    metric_type: distinct_count
+    resource: iris.dcs_iris.sepal_length
+    filters:
+      where: "species = 'versicolor'"
