+++ ui/src/components/Tabs/index.tsx
@@ -1,0 +1,2 @@
+import { Tabs } from "./Tabs";
+export default Tabs;
