--- ui/src/components/UI/Card/index.tsx
@@ -1,19 +1,0 @@
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardSubtitle,
-  CardTitle,
-  CardSection,
-} from "./Card";
-
-export {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardSubtitle,
-  CardTitle,
-  CardSection,
-};
