--- ui/src/components/UI/Piechart/index.tsx
@@ -1,2 +1,0 @@
-import { PieChart } from "./Piechart";
-export default PieChart;
